requirements.txt
pytest==8.4.1
selene==2.0.0rc9
requests==2.32.4
allure-pytest==2.15.0
appium-python-client==5.1.3
python-dotenv==1.1.1
Faker==37.5.3